Fix compiler warning

+++ src/leveldb/db/db_bench.cc
@@ -155,9 +155,9 @@ class RandomGenerator {
   }
 
   Slice Generate(int len) {
-    if (pos_ + len > data_.size()) {
+    if (pos_ + len > static_cast<int>(data_.size())) {
       pos_ = 0;
-      assert(len < data_.size());
+      assert(len < static_cast<int>(data_.size()));
     }
     pos_ += len;
     return Slice(data_.data() + pos_ - len, len);
@@ -165,11 +165,11 @@ class RandomGenerator {
 };
 
 static Slice TrimSpace(Slice s) {
-  int start = 0;
+  uint32_t start = 0;
   while (start < s.size() && isspace(s[start])) {
     start++;
   }
-  int limit = s.size();
+  uint32_t limit = s.size();
   while (limit > start && isspace(s[limit-1])) {
     limit--;
   }
@@ -425,7 +425,7 @@ class Benchmark {
     heap_counter_(0) {
     std::vector<std::string> files;
     Env::Default()->GetChildren(FLAGS_db, &files);
-    for (int i = 0; i < files.size(); i++) {
+    for (uint32_t i = 0; i < files.size(); i++) {
       if (Slice(files[i]).starts_with("heap-")) {
         Env::Default()->DeleteFile(std::string(FLAGS_db) + "/" + files[i]);
       }
